requirements.txt
markdown
pillow
django-crispy-forms
crispy-bootstrap5 
django-widget-tweaks
